requirements.txt
flask
mysql-connector-python
cryptography
python-dotenv
